# This is a combination of 10 commits. # The first commit's message is: Fix make updatedeps (hopefully)
go get in make updatedeps causes git to change packer back to the master branch. This causes CI builds in Travis to build the master branch even on pull-requests, resulting in invalid test results. This includes some additional logic to try to work around go get's odd behavior.

- Rename updatedeps target to deps so it's easier to type
- Add verifysha target will abort the run if it detects drift
- Add additional targets for tasks that commonly wrap deps so they can be run in a single command, allowing us to use make variables to track the sha
    - make ci
    - make release
- Remove previous debug output

# This is the 2nd commit message:

make updatedeps is backwards-compatible and shows a notice

# This is the 3rd commit message:

Additional checks

- Remove submake to make verifysha work correctly
- Add backwards compatibility for make updatedeps
- Guard against releasing dev builds

# This is the 4th commit message:

Abort the deps operation if the git working tree is dirty

# This is the 5th commit message:

Try to get the branch name so we don't end up with a detached head; also tweak messaging

# This is the 6th commit message:

Added missing semicolon

# This is the 7th commit message:

Hello? Yes this is packer.

# This is the 8th commit message:

Changed erroneous use of GITSHA to GITBRANCH, which makes things work in dev

# This is the 9th commit message:

Cleanup formatting and remove extra whitespace from output

# This is the 10th commit message:

Fix verifysha when we're not on a branch